purge engine_traits from file_system
