+++ components/engine_traits/src/file_system.rs
@@ -1,6 +1,43 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
+use std::sync::Arc;
+
+use file_system::{get_io_rate_limiter, get_io_type, IOOp, IORateLimiter};
+
 pub trait FileSystemInspector: Sync + Send {
     fn read(&self, len: usize) -> Result<usize, String>;
     fn write(&self, len: usize) -> Result<usize, String>;
+}
+
+pub struct EngineFileSystemInspector {
+    limiter: Option<Arc<IORateLimiter>>,
+}
+
+impl EngineFileSystemInspector {
+    #[allow(dead_code)]
+    pub fn new() -> Self {
+        EngineFileSystemInspector {
+            limiter: get_io_rate_limiter(),
+        }
+    }
+}
+
+impl FileSystemInspector for EngineFileSystemInspector {
+    fn read(&self, len: usize) -> Result<usize, String> {
+        if let Some(limiter) = &self.limiter {
+            let io_type = get_io_type();
+            Ok(limiter.request(io_type, IOOp::Read, len))
+        } else {
+            Ok(len)
+        }
+    }
+
+    fn write(&self, len: usize) -> Result<usize, String> {
+        if let Some(limiter) = &self.limiter {
+            let io_type = get_io_type();
+            Ok(limiter.request(io_type, IOOp::Write, len))
+        } else {
+            Ok(len)
+        }
+    }
 }
